# docs/swagger.json
{
    "swagger": "2.0",
    "info": {
        "description": "REST API for support ticket SLA processing system.",
        "title": "Support Ticket SLA Processing System API",
        "contact": {},
        "version": "1.0"
    },
    "host": "localhost:8080",
    "basePath": "/api/v1",
    "paths": {
        "/auth/login": {
            "post": {
                "description": "Login with username and password through Keycloak",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "Auth"
                ],
                "summary": "Login",
                "parameters": [
                    {
                        "description": "Login request",
                        "name": "request",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/request.LoginRequest"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Login successfully",
                        "schema": {
                            "$ref": "#/definitions/common.SuccessResponseDoc"
                        }
                    },
                    "400": {
                        "description": "Invalid request body",
                        "schema": {
                            "$ref": "#/definitions/common.ErrorResponseDoc"
                        }
                    },
                    "401": {
                        "description": "Invalid username or password",
                        "schema": {
                            "$ref": "#/definitions/common.ErrorResponseDoc"
                        }
                    }
                }
            }
        },
        "/reports/daily": {
            "get": {
                "security": [
                    {
                        "BearerAuth": []
                    }
                ],
                "description": "Get daily ticket report by date",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "Reports"
                ],
                "summary": "Get daily report",
                "parameters": [
                    {
                        "type": "string",
                        "description": "Report date in YYYY-MM-DD format",
                        "name": "date",
                        "in": "query"
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Get daily report successfully",
                        "schema": {
                            "$ref": "#/definitions/common.SuccessResponseDoc"
                        }
                    },
                    "400": {
                        "description": "Invalid date format",
                        "schema": {
                            "$ref": "#/definitions/common.ErrorResponseDoc"
                        }
                    },
                    "401": {
                        "description": "Unauthorized",
                        "schema": {
                            "$ref": "#/definitions/common.ErrorResponseDoc"
                        }
                    },
                    "404": {
                        "description": "Report not found",
                        "schema": {
                            "$ref": "#/definitions/common.ErrorResponseDoc"
                        }
                    },
                    "500": {
                        "description": "Internal server error",
                        "schema": {
                            "$ref": "#/definitions/common.ErrorResponseDoc"
                        }
                    }
                }
            }
        },
        "/ticket-events/import": {
            "post": {
                "security": [
                    {
                        "BearerAuth": []
                    }
                ],
                "description": "Import ticket events in batch. Accepts a multipart file upload (CSV or JSON) via the `file` field, or a raw JSON body.",
                "consumes": [
                    "multipart/form-data",
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "ticket-events"
                ],
                "summary": "Import ticket events",
                "parameters": [
                    {
                        "type": "file",
                        "description": "CSV or JSON file to import",
                        "name": "file",
                        "in": "formData"
                    },
                    {
                        "description": "Raw JSON array (when not using file upload)",
                        "name": "request",
                        "in": "body",
                        "schema": {
                            "type": "array",
                            "items": {
                                "type": "object",
                                "additionalProperties": true
                            }
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Import ticket events successfully",
                        "schema": {
                            "$ref": "#/definitions/common.SuccessResponseDoc"
                        }
                    },
                    "400": {
                        "description": "Invalid import input",
                        "schema": {
                            "$ref": "#/definitions/common.ErrorResponseDoc"
                        }
                    },
                    "401": {
                        "description": "Unauthorized",
                        "schema": {
                            "$ref": "#/definitions/common.ErrorResponseDoc"
                        }
                    },
                    "500": {
                        "description": "Internal server error",
                        "schema": {
                            "$ref": "#/definitions/common.ErrorResponseDoc"
                        }
                    }
                }
            }
        },
        "/tickets": {
            "get": {
                "security": [
                    {
                        "BearerAuth": []
                    }
                ],
                "description": "Get tickets with filters and pagination",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "Tickets"
                ],
                "summary": "List tickets",
                "parameters": [
                    {
                        "type": "string",
                        "description": "Filter by ticket status",
                        "name": "status",
                        "in": "query"
                    },
                    {
                        "type": "string",
                        "description": "Filter by priority",
                        "name": "priority",
                        "in": "query"
                    },
                    {
                        "type": "string",
                        "description": "Filter by assignee ID",
                        "name": "assignee_id",
                        "in": "query"
                    },
                    {
                        "type": "integer",
                        "description": "Page number",
                        "name": "page",
                        "in": "query"
                    },
                    {
                        "type": "integer",
                        "description": "Number of items per page",
                        "name": "limit",
                        "in": "query"
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Get tickets successfully",
                        "schema": {
                            "$ref": "#/definitions/common.SuccessResponseDoc"
                        }
                    },
                    "400": {
                        "description": "Invalid query parameters",
                        "schema": {
                            "$ref": "#/definitions/common.ErrorResponseDoc"
                        }
                    },
                    "401": {
                        "description": "Unauthorized",
                        "schema": {
                            "$ref": "#/definitions/common.ErrorResponseDoc"
                        }
                    },
                    "500": {
                        "description": "Internal server error",
                        "schema": {
                            "$ref": "#/definitions/common.ErrorResponseDoc"
                        }
                    }
                }
            },
            "post": {
                "security": [
                    {
                        "BearerAuth": []
                    }
                ],
                "description": "Create a new support ticket",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "Tickets"
                ],
                "summary": "Create ticket",
                "parameters": [
                    {
                        "description": "Create ticket request",
                        "name": "request",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/request.CreateTicketReq"
                        }
                    }
                ],
                "responses": {
                    "201": {
                        "description": "Create ticket successfully",
                        "schema": {
                            "$ref": "#/definitions/common.SuccessResponseDoc"
                        }
                    },
                    "400": {
                        "description": "Invalid request body or invalid priority",
                        "schema": {
                            "$ref": "#/definitions/common.ErrorResponseDoc"
                        }
                    },
                    "401": {
                        "description": "Unauthorized",
                        "schema": {
                            "$ref": "#/definitions/common.ErrorResponseDoc"
                        }
                    },
                    "500": {
                        "description": "Internal server error",
                        "schema": {
                            "$ref": "#/definitions/common.ErrorResponseDoc"
                        }
                    }
                }
            }
        },
        "/tickets/{id}": {
            "get": {
                "security": [
                    {
                        "BearerAuth": []
                    }
                ],
                "description": "Get ticket detail by ID",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "Tickets"
                ],
                "summary": "Get ticket detail",
                "parameters": [
                    {
                        "type": "integer",
                        "description": "Ticket ID",
                        "name": "id",
                        "in": "path",
                        "required": true
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Get ticket successfully",
                        "schema": {
                            "$ref": "#/definitions/common.SuccessResponseDoc"
                        }
                    },
                    "400": {
                        "description": "Invalid ticket ID",
                        "schema": {
                            "$ref": "#/definitions/common.ErrorResponseDoc"
                        }
                    },
                    "401": {
                        "description": "Unauthorized",
                        "schema": {
                            "$ref": "#/definitions/common.ErrorResponseDoc"
                        }
                    },
                    "404": {
                        "description": "Ticket not found",
                        "schema": {
                            "$ref": "#/definitions/common.ErrorResponseDoc"
                        }
                    },
                    "500": {
                        "description": "Internal server error",
                        "schema": {
                            "$ref": "#/definitions/common.ErrorResponseDoc"
                        }
                    }
                }
            }
        },
        "/tickets/{id}/status": {
            "patch": {
                "security": [
                    {
                        "BearerAuth": []
                    }
                ],
                "description": "Update status of a ticket by ID",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "Tickets"
                ],
                "summary": "Update ticket status",
                "parameters": [
                    {
                        "type": "integer",
                        "description": "Ticket ID",
                        "name": "id",
                        "in": "path",
                        "required": true
                    },
                    {
                        "description": "Update status request",
                        "name": "request",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/request.UpdateStatusReq"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Ticket status updated successfully",
                        "schema": {
                            "$ref": "#/definitions/common.SuccessMessageResponseDoc"
                        }
                    },
                    "400": {
                        "description": "Invalid request body or invalid status transition",
                        "schema": {
                            "$ref": "#/definitions/common.ErrorResponseDoc"
                        }
                    },
                    "401": {
                        "description": "Unauthorized",
                        "schema": {
                            "$ref": "#/definitions/common.ErrorResponseDoc"
                        }
                    },
                    "404": {
                        "description": "Ticket not found",
                        "schema": {
                            "$ref": "#/definitions/common.ErrorResponseDoc"
                        }
                    },
                    "500": {
                        "description": "Internal server error",
                        "schema": {
                            "$ref": "#/definitions/common.ErrorResponseDoc"
                        }
                    }
                }
            }
        }
    },
    "definitions": {
        "common.ErrorDetailDoc": {
            "type": "object",
            "properties": {
                "field": {
                    "type": "string",
                    "example": "field_name"
                },
                "message": {
                    "type": "string",
                    "example": "validation error message"
                }
            }
        },
        "common.ErrorDoc": {
            "type": "object",
            "properties": {
                "code": {
                    "type": "string",
                    "example": "ERROR_CODE"
                },
                "details": {
                    "type": "array",
                    "items": {
                        "$ref": "#/definitions/common.ErrorDetailDoc"
                    }
                },
                "message": {
                    "type": "string",
                    "example": "error message"
                },
                "status": {
                    "type": "integer",
                    "example": 400
                }
            }
        },
        "common.ErrorResponseDoc": {
            "type": "object",
            "properties": {
                "error": {
                    "$ref": "#/definitions/common.ErrorDoc"
                },
                "success": {
                    "type": "boolean",
                    "example": false
                }
            }
        },
        "common.SuccessMessageResponseDoc": {
            "type": "object",
            "properties": {
                "message": {
                    "type": "string",
                    "example": "operation completed successfully"
                },
                "success": {
                    "type": "boolean",
                    "example": true
                }
            }
        },
        "common.SuccessResponseDoc": {
            "type": "object",
            "properties": {
                "data": {},
                "message": {
                    "type": "string",
                    "example": "success"
                },
                "success": {
                    "type": "boolean",
                    "example": true
                }
            }
        },
        "domain.Priority": {
            "type": "string",
            "enum": [
                "low",
                "medium",
                "high"
            ],
            "x-enum-varnames": [
                "PriorityLow",
                "PriorityMedium",
                "PriorityHigh"
            ]
        },
        "domain.TicketStatus": {
            "type": "string",
            "enum": [
                "new",
                "assigned",
                "in_progress",
                "resolved",
                "closed",
                "cancelled"
            ],
            "x-enum-varnames": [
                "StatusNew",
                "StatusAssigned",
                "StatusInProgress",
                "StatusResolved",
                "StatusClosed",
                "StatusCancelled"
            ]
        },
        "request.CreateTicketReq": {
            "type": "object",
            "properties": {
                "description": {
                    "type": "string"
                },
                "priority": {
                    "$ref": "#/definitions/domain.Priority"
                },
                "sla_due_at": {
                    "type": "string"
                },
                "title": {
                    "type": "string"
                }
            }
        },
        "request.LoginRequest": {
            "type": "object",
            "required": [
                "password",
                "username"
            ],
            "properties": {
                "password": {
                    "type": "string"
                },
                "username": {
                    "type": "string"
                }
            }
        },
        "request.UpdateStatusReq": {
            "type": "object",
            "properties": {
                "note": {
                    "type": "string"
                },
                "status": {
                    "$ref": "#/definitions/domain.TicketStatus"
                }
            }
        }
    },
    "securityDefinitions": {
        "BearerAuth": {
            "type": "apiKey",
            "name": "Authorization",
            "in": "header"
        }
    }
}